+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fix: Small adjustments to Playwright tests
 - Upgrade testing-library packages to v14 and jest to v29
+- Replace fireEvent with userEvent
 
 ## [0.17.4]